requirements.txt
grpcio==1.74.0
protobuf==6.32.0
setuptools==80.9.0
grpcio-tools==1.74.0
torch==2.8.0
requests==2.32.5
python-dotenv==1.1.1
docling==2.61.2